Add isrecursive feature and more tests. bump to v1.8
