+++ __tests__/main.test.ts
@@ -1,9 +1,23 @@
 import * as main from '../src/main';
+import * as helpers from '../src/methods-helpers';
+import * as azure from '../src/methods-azure';
 
-test('validate connection_string', async () => {
-  await expect(main.uploadToAzure('', '', 'src/abcd', '', false, false)).rejects.toThrow('The connection_string cannot be empty.');
+// Make sure the error occurs when the connection string parameter is missing
+test('Azure - UploadToAzure connection_string validation', async () => {
+  await expect(azure.UploadToAzure('', '', 'src/abcd', '', false, false, true)).rejects.toThrow('The connection_string cannot be empty.');
 });
 
-test('validate source_folder', async () => {
-  await expect(main.uploadToAzure('XXXXXXXX', '', '', '', false, false)).rejects.toThrow('The source_folder was not a valid value.');
+// Make sure there is an error when the source older parameter is missing
+test('Azure - UploadToAzure source_folder validation', async () => {
+  await expect(azure.UploadToAzure('XXXXXXXX', '', '', '', false, false, true)).rejects.toThrow('The source_folder was not a valid value.');
+});
+
+test('Helpers - FindFilesFlat', async () => {
+  const files = await helpers.FindFilesFlat('./src/TestData/TextFileTestRoot');
+  expect(files.length).toBeGreaterThanOrEqual(2);
+});
+
+test('Helpers - FindFilesRecursive', async () => {
+  const files = await helpers.FindFilesRecursive('./src/TestData/');
+  expect(files.length).toBeGreaterThan(0);
 });
