+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "action-azure-upload-blob",
-  "version": "1.7.1",
+  "version": "1.8.0",
   "private": false,
   "description": "GitHub Acton that uploads file to Azure Storage blob container.",
   "main": "lib/main.js",
@@ -14,7 +14,7 @@
     "format": "prettier --write **/*.ts",
     "lint": "eslint src/**/*.ts",
     "package": "ncc build --source-map --license licenses.txt",
-    "test": "jest",
+    "test": "jest --coverage",
     "pre-commit": "npm run rebuild && npm run format && npm run lint && npm run package",
     "all": "npm run rebuild && npm run format && npm run lint && npm run package && npm test"
   },
